Refactor Transfer page/components
